Split out tryAndCustomizeTheme
